Code refactoring to generator
